+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.2.0 - 2019.08.26]
+### Changed
+- Added default value 'QUERY' for query language
+
 ## [3.1.0 - 2019-07-02]
 ### Changed
 - Cleaner code